requirements.txt
streamlit
yfinance
requests
pandas
vaderSentiment
alpaca-py
plotly